refactor:레포지토리 / 서비스 클래스명 변경
