+++ src/main/java/gift/repository/userRepository/UserRepository.java
@@ -1,23 +1,18 @@
-//package gift.repository.userRepository;
-//
-//import gift.entity.User;
-//import gift.entity.UserRole;
-//
-//import java.util.List;
-//
-//public interface UserRepository {
-//    User save(String email, String password, UserRole role);
-//
-//    List<User> getAllUsers();
-//
-//    List<User> findUsersByEmail(String email);
-//
-//    User findUserById(Long id);
-//
-//    User updateUser(User findUser, String email, String password);
-//
-//    void deleteUser(User findUser);
-//
-//    User findUserByEmail(String email);
-//
-//}
+package gift.repository.userRepository;
+
+import gift.entity.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface UserRepository extends JpaRepository<User, Long> {
+    User findUserById(Long id);
+
+    User findUserByEmail(String email);
+
+    void deleteUserById(Long id);
+
+    boolean findEmailByEmail(String email);
+
+    boolean existsByEmail(String email);
+}
