+++ src/main/java/gift/repository/wishListRepository/WishListRepository.java
@@ -1,16 +1,19 @@
-//package gift.repository.wishListRepository;
-//
-//import gift.entity.WishItem;
-//
-//import java.util.List;
-//
-//public interface WishListRepository {
-//
-//    WishItem addWishItem(Long userId, Long itemId, Integer quantity);
-//
-//    List<WishItem> getAllWishItems(Long userId);
-//
-//    WishItem updateWishItem(Integer quantity, Long itemId, Long userId);
-//
-//    WishItem deleteWishItem(Long userId, Long itemId);
-//}
+package gift.repository.wishListRepository;
+
+import gift.entity.Item;
+import gift.entity.User;
+import gift.entity.WishItem;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import java.util.List;
+import java.util.Optional;
+
+@Repository
+public interface WishListRepository extends JpaRepository<WishItem, Long> {
+    List<WishItem> findAllByUser(User user);
+
+    Optional<WishItem> findByUserAndItem(User user, Item item);
+
+    boolean existsByItem(Item item);
+}
