+++ src/main/java/gift/repository/wishListRepository/WishListRepositorySub.java
@@ -1,0 +1,16 @@
+//package gift.repository.wishListRepository;
+//
+//import gift.entity.WishItem;
+//
+//import java.util.List;
+//
+//public interface WishListRepositorySub {
+//
+//    WishItem addWishItem(Long userId, Long itemId, Integer quantity);
+//
+//    List<WishItem> getAllWishItems(Long userId);
+//
+//    WishItem updateWishItem(Integer quantity, Long itemId, Long userId);
+//
+//    WishItem deleteWishItem(Long userId, Long itemId);
+//}
